Replace server.js with server/main.ts
