+++ server/main.ts
@@ -1,0 +1,35 @@
+import * as bodyParser from 'body-parser';
+import * as express from 'express';
+import * as mongoose from 'mongoose';
+import * as path from 'path';
+
+const MONGO_URL = `mongodb://cerberus:${process.env.MLAB_PASSWORD}@ds127802.mlab.com:27802/heroku_vhdmcrh0`;
+
+const app = express();
+app.set('port', (process.env.PORT || 3000));
+
+app.use('/', express.static(path.join(__dirname, '../public')));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+
+mongoose.connect(MONGO_URL);
+const db = mongoose.connection;
+(<any>mongoose).Promise = global.Promise;
+
+db.on('error', console.error.bind(console, 'connection error:'));
+db.once('open', () => {
+  console.log('Connected to MongoDB');
+
+  // Set routes here
+
+  app.get('*', (req, res) => {
+    res.sendFile(path.join(__dirname, '../dist/index.html'));
+  });
+
+  app.listen(app.get('port'), () => {
+    console.log('Angular Full Stack listening on port ' + app.get('port'));
+  });
+
+});
+
+export { app };
